Finish gateway order condition
