+++ src/base/Gateway.php
@@ -8,15 +8,10 @@
 namespace craft\commerce\base;
 
 use Craft;
-use craft\base\ElementInterface;
 use craft\base\SavableComponent;
-use craft\commerce\elements\conditions\orders\DiscountOrderCondition;
-use craft\commerce\elements\conditions\orders\GatewayOrderCondition;
 use craft\commerce\elements\Order;
 use craft\commerce\models\payments\BasePaymentForm;
 use craft\commerce\models\Transaction;
-use craft\elements\conditions\ElementConditionInterface;
-use craft\helpers\Json;
 use craft\helpers\StringHelper;
 use craft\helpers\UrlHelper;
 
@@ -125,25 +120,18 @@ abstract class Gateway extends SavableComponent implements GatewayInterface
     }
 
     /**
-     * @var ElementConditionInterface|null
-     */
-    private ?ElementConditionInterface $_orderCondition = null;
-
-    /**
      * @inheritdoc
      */
     public function availableForUseWithOrder(Order $order): bool
     {
-        // First check if the gateway has an order condition
-        if ($this->hasOrderCondition()) {
-            return $this->getOrderCondition()->matchElement($order);
-        } 
-        
-        // Fall back to the deprecated frontend enabled setting
-        if (method_exists($this, 'getIsFrontendEnabled')) {
-            return $this->getIsFrontendEnabled();
+        if (Craft::$app->getRequest()->getIsSiteRequest() && !$this->getIsFrontendEnabled()) {
+            return false;
         }
-        
+
+        if ($this->hasOrderCondition() && !$this->getOrderCondition()->matchElement($order)) {
+            return false;
+        }
+
         return true;
     }
 
@@ -157,53 +145,12 @@ abstract class Gateway extends SavableComponent implements GatewayInterface
     
     /**
      * Returns true if this gateway has an order condition
-     * 
+     *
      * @since 5.3.5
      */
     public function hasOrderCondition(): bool
     {
         return $this->getOrderCondition()->getConditionRules() !== [];
-    }
-    
-    /**
-     * Gets the order condition for this gateway
-     * 
-     * @since 5.4.0
-     */
-    public function getOrderCondition(): ElementConditionInterface
-    {
-        /** @var DiscountOrderCondition $condition */
-        $condition = $this->_orderCondition ?? new GatewayOrderCondition();
-        $condition->mainTag = 'div';
-        $condition->name = 'orderCondition';
-
-        return $condition;
-    }
-    
-    /**
-     * Sets the order condition for this gateway
-     * 
-     * @since 5.3.5
-     */
-    public function setOrderCondition(ElementConditionInterface|string|array $condition): void
-    {
-        if (empty($condition)) {
-            $this->_orderCondition = null;
-            return;
-        }
-
-        if (is_string($condition)) {
-            $condition = Json::decodeIfJson($condition);
-        }
-
-        if (!$condition instanceof DiscountOrderCondition) {
-            $condition['class'] = DiscountOrderCondition::class;
-            /** @var DiscountOrderCondition $condition */
-            $condition = Craft::$app->getConditions()->createCondition($condition);
-        }
-        $condition->forProjectConfig = true;
-
-        $this->_orderCondition = $condition;
     }
 
     /**
