+++ src/elements/conditions/orders/GatewayOrderCondition.php
@@ -8,26 +8,13 @@
 namespace craft\commerce\elements\conditions\orders;
 
 use craft\commerce\elements\Order;
-use craft\elements\conditions\ElementCondition;
 
 /**
  * Gateway Order Condition
  *
  * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
- * @since 5.3.5
+ * @since 5.4.0
  */
 class GatewayOrderCondition extends OrderCondition
 {
-    /**
-     * @inheritdoc
-     */
-    protected function conditionRuleTypes(): array
-    {
-        return array_merge(
-            parent::conditionRuleTypes(),
-            [
-                EnabledGatewayConditionRule::class,
-            ]
-        );
-    }
 }
