+++ src/records/Gateway.php
@@ -20,6 +20,7 @@ use DateTime;
  * @property int $id
  * @property bool $isArchived
  * @property string $name
+ * @property string $orderCondition
  * @property string $paymentType
  * @property array $settings
  * @property int $sortOrder
