+++ src/templates/settings/gateways/_edit.twig
@@ -98,6 +98,17 @@
         </div>
     {% endfor %}
 
+    {{ forms.booleanMenuField({
+        label: "Enabled for customers to select during checkout?"|t('commerce'),
+        id: 'isFrontendEnabled',
+        name: 'isFrontendEnabled',
+        includeEnvVars: true,
+        value: gateway.isFrontendEnabled(false),
+        errors: gateway.getErrors('isFrontendEnabled'),
+        disabled: readOnly,
+    }) }}
+
+
     <hr>
     {% set orderConditionInput %}
         {{ gateway.getOrderCondition().getBuilderHtml()|raw }}
